Refactor into telugu_agents and removed changes to avatar.py and put the telugu avatars in telugu_agents.py
