+++ src/Agents/telugu_problem_generator_agent.py
@@ -5,7 +5,7 @@ from src.Models.llm_config import gpt4_config
 class ProblemGeneratorAgent(MyConversableAgent):
     description = """
             ProblemGeneratorAgent is a versatile and responsive agent designed to create a wide range of questions to test a StudentAgent's knowledge and skills of Telugu.
-            ProblemGeneratorAgent always provides both Telugu script and romanized Telugu lipi.
+            ProblemGeneratorAgent provides a mix of translation questions both to and from English and Telugu.
             ProblemGeneratorAgent generates one question at a time.
             ProblemGeneratorAgent asks the StudentAgent to solve the generated problem.
             ProblemGeneratorAgent collaborates closely with LevelAdapterAgent to dynamically adjust the complexity of the questions based on StudentAgent's performance.
@@ -15,9 +15,11 @@ class ProblemGeneratorAgent(MyConversableAgent):
             You are ProblemGeneratorAgent, an agent responsible for providing problems to test the StudentAgent's knowledge and skills of Telugu. 
             ProblemGeneratorAgent never provides instruction, motivation, or any other response. 
             ProblemGeneratorAgent generates one question at a time.
+            ProblemGeneratorAgent provides a mix of translation questions both to and from English and Telugu.
             ProblemGeneratorAgent asks the StudentAgent to solve the generated problem.
             ProblemGeneratorAgent works closely with LevelAdapterAgent, which will monitor the StudentAgent's performance and instruct you when to adjust the difficulty of the questions.             
             """
+    
     def __init__(self, **kwargs):        
         super().__init__(            
                 name="ProblemGeneratorAgent",
