+++ src/UI/avatar.py
@@ -1,5 +1,5 @@
 from src.Agents.agents import *
-from Agents.telugu_agents import telugu_teaching_agent
+
 avatar = {
     student.name: "✏️",                 # Pencil
     knowledge_tracer.name: "🧠",       # Brain
@@ -13,8 +13,7 @@ avatar = {
     level_adapter.name: "📈",      # Chart with upwards trend for level adaptation
     motivator.name: "🏆",  
     gamification.name: "🎮",        # Trophy emoji for motivation
-    telugu_teaching_agent.name: "🎮",
-}
+ }
 
 avatar2 = {
     "Learner": "🎓",        # Graduate cap: Represents the learner
